extract definition and phony checker hook
